@@ -13,17 +13,17 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^6.0.3",
+    "@angular/animations": "6.0.4",
     "@angular/cdk": "^6.2.1",
-    "@angular/common": "^5.0.0",
-    "@angular/compiler": "^5.0.0",
-    "@angular/core": "^5.0.0",
-    "@angular/forms": "^5.0.0",
-    "@angular/http": "^5.0.0",
+    "@angular/common": "6.0.4",
+    "@angular/compiler": "6.0.4",
+    "@angular/core": "6.0.4",
+    "@angular/forms": "6.0.4",
+    "@angular/http": "6.0.4",
     "@angular/material": "^6.2.1",
-    "@angular/platform-browser": "^5.0.0",
-    "@angular/platform-browser-dynamic": "^5.0.0",
-    "@angular/router": "^5.0.0",
+    "@angular/platform-browser": "6.0.4",
+    "@angular/platform-browser-dynamic": "6.0.4",
+    "@angular/router": "6.0.4",
     "@compodoc/compodoc": "^1.1.3",
     "bootstrap": "^3.3.7",
     "chart.js": "^2.7.1",
@@ -33,14 +33,16 @@
     "ng": "0.0.0-rc6",
     "ng2-charts": "^1.6.0",
     "ng2-order-pipe": "^0.1.5",
-    "rxjs": "^5.5.2",
-    "zone.js": "^0.8.14"
+    "rxjs": "^6.2.1",
+    "rxjs-compat": "^6.2.1",
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular-devkit/core": "0.0.29",
-    "@angular/cli": "1.6.4",
-    "@angular/compiler-cli": "^5.0.0",
-    "@angular/language-service": "^5.0.0",
+    "@angular-devkit/build-angular": "~0.6.8",
+    "@angular-devkit/core": "0.6.8",
+    "@angular/cli": "6.0.8",
+    "@angular/compiler-cli": "6.0.4",
+    "@angular/language-service": "6.0.4",
     "@types/jasmine": "~2.8.4",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "~9.3.0",
@@ -59,7 +61,7 @@
     "protractor": "~5.2.2",
     "protractor-html-reporter": "^1.3.2",
     "ts-node": "~4.1.0",
-    "tslint": "~5.9.1",
-    "typescript": "^2.5.3"
+    "tslint": "^5.10.0",
+    "typescript": "2.7.2"
   }
 }
