@@ -1,7 +1,7 @@
 import { Injectable } from "@angular/core";
 import { HttpClient } from '@angular/common/http';
-import { Observable } from "rxjs/Observable";
-import 'rxjs/add/operator/map'
+import { Observable } from "rxjs";
+
 import { Associate } from "../../models/associate.model";
 import { Response } from "@angular/http/";
 import { environment } from "../../../environments/environment";
