@@ -3,7 +3,7 @@
 import {Injectable} from '@angular/core';
 import {environment} from '../../../environments/environment';
 import {HttpClient} from '@angular/common/http';
-import {Observable} from 'rxjs/Observable';
+import {Observable} from 'rxjs';
 import {Batch} from '../../models/batch.model';
 import {Associate} from '../../models/associate.model';
 
