@@ -2,9 +2,8 @@ import { Component, OnInit } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { RequestService } from '../../services/request-service/request.service';
 import { ClientService } from '../../services/client-service/client.service';
-import { Subject } from 'rxjs/Subject';
+// import { Subject ,  Observable } from 'rxjs';
 import { Client } from '../../models/client.model';
-import { Observable } from 'rxjs/Observable';
 import { SelectedStatusConstants } from '../../constants/selected-status.constants';
 import { ThemeConstants } from '../../constants/theme.constants';
 import { Color } from 'ng2-charts';
