@@ -32,7 +32,7 @@
  
  
  // Only implements params and part of snapshot.paramMap
- import { BehaviorSubject } from 'rxjs/BehaviorSubject';
+ import { BehaviorSubject } from 'rxjs';
  import { convertToParamMap, ParamMap } from '@angular/router';
  
  @Injectable()
